We are now actually scrapping items
The UI check is most likely being executed for other UI as well; I need to add a check whether the UI opened is a scrapper or not.

I also need to somehow work around the orig call, as that shows the UI with the items that were already scrapped.